requirements.txt
certifi==2023.5.7
dnspython==2.4.2
h2==4.1.0
